Add user conversations fan out
Added fan out for user conversations in Firebase to retirieve only conversations between the two users who are involved.

Added a setupNameAndAvatar() function that fetches the other contact's name and avatar to display in your conversation.

Fixes issue #2 with the setupUserConversations() function. This clears the old messages array if a new user logs in on the same device.