Add chat bubbles for sender
Added chat bubble UI for sent messages.

Fixed an issue in observeUserMessages. When a new message was sent from the recipient in a conversation, their message would create a new conversation. To fix this I changed the observation value from toId to chatPartnerId, referencing the chatPartnerId function inside of the Message model